signup password confirm

+++ testdb.php
@@ -44,12 +44,12 @@
 		$testRes = $dao->getLogin($user,$pass);
 		echo print_r($testRes->fetch(PDO::FETCH_ASSOC));
 		//echo print("SELECT * FROM Users WHERE username='{$user}' AND password='{$pass}'");*/
-		/*$testAdd = $dao->createUser('testu','password','email.email');
-		echo print($testAdd);*/
-		/*if($testAdd){*/
+		$testAdd = $dao->createUser('testu','password','email.email');
+		echo print($testAdd);
+		if($testAdd){
 		  $testDelete = $dao->deleteUser('testu','password');
 		  echo print($testDelete);
-		/*}*/
+		}
 	?>
 	</p>
 	</body>
